requirements.txt
python-telegram-bot==20.7
pandas==1.5.3
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3